Final Commit
Fleshed out the gameStatus function as well as added a few more functions to check for a win, check if the board was full, and to print out the status. The print function doubles as a bool that allowed me to do some logic to automatically end the game when a win condition was met.