CPP & Header Second Commit
Put the enums in the header class in the public scope for use in my if statements in the main.cpp. Also added in a couple functions, resetBoard and checkStatus. Reset board is used for looping purposes so when you end a game it will reset and allow you to play again. checkStatus changes the game status accordingly throughout the game.